Fix for #41 and a start at #5

+++ Imagr/Utils.py
@@ -221,6 +221,12 @@ def downloadAndInstallPackage(url, target, progress_method=None):
         # Clean up after ourselves
         shutil.rmtree(temp_dir)
 
+def runScript(script, target, progress_method=None):
+    """
+    Runs a script on a specific volume
+    """
+    NSLog("Running script on %@", target)
+    NSLog("Script: %@", script)
 
 def installPkg(pkg, target, progress_method=None):
     """
@@ -385,46 +391,54 @@ def copyFirstBoot(root):
     firstboot_dir = 'usr/local/first-boot'
     plistlib.writePlist(config_plist, os.path.join(root, firstboot_dir, 'config.plist'))
 
-    # Copy the LaunchDaemon, LaunchAgent and LoginLog.app to the right places
+    # Copy the LaunchDaemon, LaunchAgent and Log.app to the right places
     script_dir = os.path.dirname(os.path.realpath(__file__))
     NSLog(str(script_dir))
     launchDaemon_dir = os.path.join(root, 'Library', 'LaunchDaemons')
     if not os.path.exists(launchDaemon_dir):
         os.makedirs(launchDaemon_dir)
-    shutil.copy(os.path.join(script_dir,
-    'com.grahamgilbert.first-boot-pkg.plist'), os.path.join(launchDaemon_dir,
-    'com.grahamgilbert.first-boot-pkg.plist'))
-    # Set the permisisons
-    os.chmod(os.path.join(launchDaemon_dir,
-    'com.grahamgilbert.first-boot-pkg.plist'), 0644)
-    os.chown(os.path.join(launchDaemon_dir,
-    'com.grahamgilbert.first-boot-pkg.plist'), 0, 0)
+
+    if not os.path.exists(os.path.join(launchDaemon_dir,
+    'com.grahamgilbert.first-boot-pkg.plist')):
+        shutil.copy(os.path.join(script_dir,
+        'com.grahamgilbert.first-boot-pkg.plist'), os.path.join(launchDaemon_dir,
+        'com.grahamgilbert.first-boot-pkg.plist'))
+        # Set the permisisons
+        os.chmod(os.path.join(launchDaemon_dir,
+        'com.grahamgilbert.first-boot-pkg.plist'), 0644)
+        os.chown(os.path.join(launchDaemon_dir,
+        'com.grahamgilbert.first-boot-pkg.plist'), 0, 0)
 
     launchAgent_dir = os.path.join(root, 'Library', 'LaunchAgents')
     if not os.path.exists(launchAgent_dir):
         os.makedirs(launchAgent_dir)
-    shutil.copy(os.path.join(script_dir, 'se.gu.it.LoginLog.plist'),
-    os.path.join(launchAgent_dir, 'se.gu.it.LoginLog.plist'))
-    # Set the permisisons
-    os.chmod(os.path.join(launchAgent_dir, 'se.gu.it.LoginLog.plist'), 0644)
-    os.chown(os.path.join(launchAgent_dir, 'se.gu.it.LoginLog.plist'), 0, 0)
+
+    if not os.path.exists(os.path.join(launchAgent_dir, 'se.gu.it.LoginLog.plist'))
+        shutil.copy(os.path.join(script_dir, 'se.gu.it.LoginLog.plist'),
+        os.path.join(launchAgent_dir, 'se.gu.it.LoginLog.plist'))
+        # Set the permisisons
+        os.chmod(os.path.join(launchAgent_dir, 'se.gu.it.LoginLog.plist'), 0644)
+        os.chown(os.path.join(launchAgent_dir, 'se.gu.it.LoginLog.plist'), 0, 0)
 
     helperTools_dir = os.path.join(root, 'Library', 'PrivilegedHelperTools')
     if not os.path.exists(helperTools_dir):
         os.makedirs(helperTools_dir)
-    shutil.copytree(os.path.join(script_dir, 'LoginLog.app'),
-    os.path.join(helperTools_dir, 'LoginLog.app'))
-    # Set the permisisons
-    for root_dir, dirs, files in os.walk(os.path.join(helperTools_dir, 'LoginLog.app')):
-      for momo in dirs:
-        os.chown(os.path.join(root_dir, momo), 0, 0)
-        os.chmod(os.path.join(root_dir, momo), 0755)
-      for momo in files:
-        os.chown(os.path.join(root_dir, momo), 0, 0)
-        os.chmod(os.path.join(root_dir, momo), 0755)
 
-    # copy the script
-    shutil.copy(os.path.join(script_dir, 'first-boot'), os.path.join(root, firstboot_dir, 'first-boot'))
-    # Set the permisisons
-    os.chmod(os.path.join(root, firstboot_dir, 'first-boot'), 0755)
-    os.chown(os.path.join(root, firstboot_dir, 'first-boot'), 0, 0)
+    if not os.path.exists(os.path.join(helperTools_dir, 'LoginLog.app')):
+        shutil.copytree(os.path.join(script_dir, 'LoginLog.app'),
+        os.path.join(helperTools_dir, 'LoginLog.app'))
+        # Set the permisisons
+        for root_dir, dirs, files in os.walk(os.path.join(helperTools_dir, 'LoginLog.app')):
+          for momo in dirs:
+            os.chown(os.path.join(root_dir, momo), 0, 0)
+            os.chmod(os.path.join(root_dir, momo), 0755)
+          for momo in files:
+            os.chown(os.path.join(root_dir, momo), 0, 0)
+            os.chmod(os.path.join(root_dir, momo), 0755)
+
+    if not os.path.exists(os.path.join(root, firstboot_dir, 'first-boot')):
+        # copy the script
+        shutil.copy(os.path.join(script_dir, 'first-boot'), os.path.join(root, firstboot_dir, 'first-boot'))
+        # Set the permisisons
+        os.chmod(os.path.join(root, firstboot_dir, 'first-boot'), 0755)
+        os.chown(os.path.join(root, firstboot_dir, 'first-boot'), 0, 0)
